Switch back to arch-based builds
The build process of pygithub makes use of 2to3 for Python 3 support.
Thus this means that noarch packages would end up as either Python 3 or
Python 2 only depending on what was used during the build. IOW the
packages would be broken for one of the Pythons. Hence drop noarch and
make a note of this for future maintainers.